+++ rootdriver/db/__init__.py
@@ -1,1 +1,2 @@
 from  .json_db import  JsonDB
+from .base_db import BaseDB