+++ rootdriver/utils/optional_dependence.py
@@ -1,0 +1,12 @@
+from importlib.metadata import version, PackageNotFoundError
+
+from ..exceptions import OptionalDependenceNotFoundError
+
+
+def deal_optional_dependence_installed_status(dependence_name: str):
+    try:
+        version(dependence_name)
+        return True
+    except PackageNotFoundError:
+        raise OptionalDependenceNotFoundError(dependence_name)
+